# legacy/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "FastAPI",
        "version": "0.1.0"
    },
    "paths": {
...
        "/settings/tab_group_main_state/{user_id}": {
            "get": {
                "tags": [
                    "settings"
                ],
                "summary": "Retrieve the tab group main state for a user",
                "description": "Retrieve the tab group main state for a user, which indicates the currently active tab.",
                "operationId": "get_tab_group_main_state_settings_tab_group_main_state__user_id__get",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "description": "Should be a valid user id that corresponds to a user in the user table",
                            "title": "User Id"
                        },
                        "description": "Should be a valid user id that corresponds to a user in the user table"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TabGroupMainStateModel"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "settings"
                ],
                "summary": "Update the tab group main state for a user",
                "description": "Update the tab group main state for a user, which indicates the currently active tab.",
                "operationId": "update_tab_group_main_state_settings_tab_group_main_state__user_id__put",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "description": "Should be a valid user id that corresponds to a user in the user table",
                            "title": "User Id"
                        },
                        "description": "Should be a valid user id that corresponds to a user in the user table"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TabGroupMainStateModel"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TabGroupMainStateModel"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "settings"
                ],
                "summary": "Delete the tab group main state for a user",
                "description": "Delete the tab group main state for a user, which indicates the currently active tab.",
                "operationId": "delete_tab_group_main_state_settings_tab_group_main_state__user_id__delete",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "description": "Should be a valid user id that corresponds to a user in the user table",
                            "title": "User Id"
                        },
                        "description": "Should be a valid user id that corresponds to a user in the user table"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful Response"
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/settings/tab_group_main_state": {
            "post": {
                "tags": [
                    "settings"
                ],
                "summary": "Create a new tab group main state for a user",
                "description": "Create a new tab group main state for a user, which indicates the currently active tab.",
                "operationId": "create_tab_group_main_state_settings_tab_group_main_state_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TabGroupMainStateModel"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TabGroupMainStateModel"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
...
            "TabGroupMainStateModel": {
                "properties": {
                    "user_id": {
                        "type": "integer",
                        "title": "User Id"
                    },
                    "which_tab": {
                        "type": "string",
                        "title": "Which Tab"
                    }
                },
                "type": "object",
                "required": [
                    "user_id",
                    "which_tab"
                ],
                "title": "TabGroupMainStateModel"
            },
...
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            },
...
        }
    },
...
}